0.5 Updated Registry
Lets say A extends B. Now registry can bind A's objects to B's sub registry. You just need to put B.class to the new parameter. If you don't it will get the class from the given object and the object will be registered to A's sub registry because it's an object of class A.